backend/requirements.txt
requests
python-dotenv
yandex-cloud-ml-sdk
yandex-ai-studio-sdk
beautifulsoup4
lxml
pydantic
fastapi
uvicorn
httpx
numpy
newspaper4k[nlp]
scrapfly-sdk
dspy-ai
redis
rank_bm25
sentence-transformers
optuna
tqdm
